Ocean Shader, Player Movement Control, Lightweight Render Pipeline and a broken Enemy AI Script because I am missing a few things
Today i worked on creating an ocean shader, adding it to the plane with the ocean component to make it look more like an ocean. I started using the Lightweight Render Pipeline instead of the default RP and I also downloaded a couple other packages. Finally, I started working on an Enemy AI script it however its not finished yet because there is another tutorial i need to watch to complete the script.