Rudder, Island, Crystal Pylons, Trail, Deleted Code, Removed Colliders and Create Colliders
Today during class i fixed the trail after showing Ash my game as it is now. After that i added a rudder to the Crystal Voyager, during the online class Ash noted the fact that the crystal pylon is at the front of the ship and that when the flag is made/fixed you wont be able to see it, after thinking about it i realised he has a point and so i added 2 extra crystal pylons to the sides of the ship making it a little more visible(most likley will delete and create better ones when i create a better Crystal Voyager). I also deleted the enemy scripts because the tutorials rely on pre existing downloadedable code from CodeMonkey which i downloaded from their website and also because Ash will teach us how to create basic enemy AI with one line of code,  Ash told me to remove all the colliders from all the game objects that make the Crystal Voyager and put a collider in a parented game object with all the Crystal Voyager game objects set as children. I created an island called 'Wantree Island'.